Fix export to support Cloudflare Pages
